backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
minio==7.2.12
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
gunicorn==23.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
